Fix: removed need for spoolss.dll
WritePrinter() is actually (also!) in the winspool dll.   MSDN documentation has an error in referring to spoolss.dll, which is actually the server-side code used by the printspooler itself, and which shouldn't be loaded by end-user applications.

also added an alias ('LIST) to the enum func, just to make it more obvious to end-user  (printer/list) as opposed to (printer/enum)